+++ v3/pkg/application/application.go
@@ -5,7 +5,6 @@ import (
 	"embed"
 	"encoding/json"
 	"fmt"
-	"github.com/wailsapp/wails/v3/internal/fileexplorer"
 	"io"
 	"log"
 	"log/slog"
@@ -16,6 +15,8 @@ import (
 	"strconv"
 	"strings"
 	"sync"
+
+	"github.com/wailsapp/wails/v3/internal/fileexplorer"
 
 	"github.com/wailsapp/wails/v3/internal/operatingsystem"
 
@@ -1046,8 +1047,8 @@ func (a *App) Paths(selector Paths) []string {
 	return pathdirs[selector]
 }
 
-func (a *App) OpenDirectory(path string) error {
+func (a *App) OpenDirectory(path string, selectFile bool) error {
 	return InvokeSyncWithError(func() error {
-		return fileexplorer.Open(path)
+		return fileexplorer.Open(path, selectFile)
 	})
 }